+++ sentry_streams/src/filter_step.rs
@@ -214,7 +214,7 @@ mod tests {
 
     #[test]
     fn test_build_filter() {
-        crate::testutils::initialize_python();
+        crate::test_operators::initialize_python();
         traced_with_gil!(|py| {
             let callable = make_lambda(py, c_str!("lambda x: 'test' in x.payload"));
             let submitted_messages = Arc::new(Mutex::new(Vec::new()));
